+++ .npmignore
@@ -6,7 +6,7 @@
 .prettierrc
 .vscode/
 example-expo/
-example-gifted-chat/
+example-easy-chat/
 example-slack-message/
 flow-typedefs/
 media/
