+++ example-easy-chat/src/InputToolbar.js
