+++ example-easy-chat/src/MessageContainer.js
