+++ flow-typedefs/EasyAvatar.js.flow
@@ -6,7 +6,7 @@ import type {
   TextStyleProp,
 } from 'react-native/Libraries/StyleSheet/StyleSheet'
 
-export type GiftedAvatarProps = $ReadOnly<{|
+export type EasyAvatarProps = $ReadOnly<{|
   user?: User,
   avatarStyle?: ImageStyleProp,
   textStyle?: TextStyleProp,
@@ -14,4 +14,4 @@ export type GiftedAvatarProps = $ReadOnly<{|
   onLongPress: (any) => void,
 |}>
 
-export default class GiftedAvatar extends Component<GiftedAvatarProps> {}
+export default class EasyAvatar extends Component<EasyAvatarProps> {}
