+++ flow-typedefs/index.js.flow
@@ -1,2 +1,2 @@
 // @flow
-export * from './GiftedChat'
+export * from './EasyChat'
