+++ src/__tests__/Actions.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { Actions } from '../GiftedChat'
+import { Actions } from '../EasyChat'
 
 it('should render <Actions /> and compare with snapshot', () => {
   const tree = renderer.create(<Actions />).toJSON()
