+++ src/__tests__/InputToolbar.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { InputToolbar } from '../GiftedChat'
+import { InputToolbar } from '../EasyChat'
 
 it('should render <InputToolbar /> and compare with snapshot', () => {
   const tree = renderer.create(<InputToolbar />).toJSON()
