+++ src/__tests__/LoadEarlier.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { LoadEarlier } from '../GiftedChat'
+import { LoadEarlier } from '../EasyChat'
 
 it('should render <LoadEarlier /> and compare with snapshot', () => {
   const tree = renderer.create(<LoadEarlier />).toJSON()
