+++ src/__tests__/MessageContainer.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { MessageContainer } from '../GiftedChat'
+import { MessageContainer } from '../EasyChat'
 
 it('should render <MessageContainer /> and compare with snapshot', () => {
   const tree = renderer.create(<MessageContainer />).toJSON()
