+++ src/__tests__/MessageImage.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { MessageImage } from '../GiftedChat'
+import { MessageImage } from '../EasyChat'
 
 it('should render <MessageImage /> and compare with snapshot', () => {
   const tree = renderer.create(<MessageImage />).toJSON()
