+++ src/__tests__/Send.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { Send } from '../GiftedChat'
+import { Send } from '../EasyChat'
 
 it('should render <Send /> and compare with snapshot', () => {
   const tree = renderer.create(<Send />).toJSON()
