+++ src/__tests__/SystemMessage.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import renderer from 'react-test-renderer'
 
-import { SystemMessage } from '../GiftedChat'
+import { SystemMessage } from '../EasyChat'
 
 it('should render <SystemMessage /> and compare with snapshot', () => {
   const tree = renderer.create(<SystemMessage />).toJSON()
