+++ src/__tests__/Time.test.tsx
@@ -2,7 +2,7 @@ import 'react-native'
 import React from 'react'
 import createComponentWithContext from './context'
 
-import { Time } from '../GiftedChat'
+import { Time } from '../EasyChat'
 
 it('should render <Time /> and compare with snapshot', () => {
   const component = createComponentWithContext(<Time />)
